Airflow/requirements.txt
pandas
kaggle
boto3
python-dotenv
apache-airflow[amazon]
